refactor: move GameMode to model package
